+++ lib/utils/imports/import_format.dart
@@ -1,0 +1,11 @@
+enum ImportFormat {
+  none,
+  dailyYouJson,
+  daybook,
+  daylio,
+  diarium,
+  diaro,
+  myBrain,
+  oneShot,
+  pixels,
+}
